refactor: AnalysisViewMode클래스를 partial로 분리
- AnalysisViewModel클래스를 partial로(상속받은 클래스의 인터페이스 때문에 partial로 되어 있기는 했지만 파일은 하나로 유지되었었음) 각 해당하는 부분으로 여러 파일로 분리 관리
- 기존에 있던 부분에서 공통적인 생성자와 전체적인 바인딩 command는 원래 기존 파일에 유지
- 그외는 각 부분에 맞게 Filter, TiemChart, PieChart, ColumnChart, Table로 분리하여 해당 부분 코드 이동
